Drastically reduced payload size
Fixed #10 - Created compression contracts that serve to reduce payload
sizes by over 80%.

In addition to reducing payloads I also added a ShipManager to both the
server and the client in order to manage the ships on the screen.  Had
one for bullets so might as well have one for ships.

Along side of that just did general formatting changes.