Modified payloads to only send deltas to clients
Now only send deltas to client.  This involved creating an altered
attribute to the collidable object in addition to caching the last
payload to see if we nee dto send new data down to the client.  Also
ended up creating a User clas sto represent each client.  Part of this
ended upmaking sure the client and server were in sync for payload
retrievals.

Fixed #26